ex 4.21* - use toString for ObjectID

+++ part4/bloglist/controllers/blogs.js
@@ -55,8 +55,37 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async(request, response, next) => {
   try {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    if (!user) {
+      return response.status(400).json({ error: 'userId missing or not valid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+
+    if(!blog) {
+      return response.status(204)
+    }
+
+    // to string is necessary because id is an object **
+    console.log(blog)
+    console.log(user)
+    if(blog.user.toString() === user.id) // do not compare blog and user's ids lol - 
+    {
+      await Blog.findByIdAndDelete(request.params.id)
+
+      user.blogs.pull(request.params.id)
+
+      await user.save()
+
+      response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+    //await Blog.findByIdAndDelete(request.params.id)
   }
   catch(error) {
     next(error)
